requirements.txt
asgiref==3.8.1
asn1crypto==1.5.1
Django==5.1.5
django-filter==24.3
django-phonenumber-field==8.0.0
django-rest-framework==0.1.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
pg8000==1.31.2
phonenumbers==8.13.54
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
scramp==1.4.5
six==1.17.0
sqlparse==0.5.3
unidecode==1.3.8
